Library

Quantum Circuit

Snowflake.QuantumCircuit — Type

    QuantumCircuit(qubit_count = .., bit_count = ...)

A data structure to represent a quantum circuit.

Fields

qubit_count::Int – number of qubits (i.e. quantum register size).

bit_count::Int – number of classical bits (i.e. classical register size).

id::UUID – a universally unique identifier for the circuit. A UUID is automatically generated once an instance is created.

pipeline::Array{Array{Gate}} – the pipeline of gates to operate on qubits.

Examples

julia> c = Snowflake.QuantumCircuit(qubit_count = 2, bit_count = 0)
Quantum Circuit Object:
   id: b2d2be56-7af2-11ec-31a6-ed9e71cb3360 
   qubit_count: 2 
   bit_count: 0 
q[1]:
     
q[2]:

Snowflake.push_gate! — Function

    push_gate!(circuit::QuantumCircuit, gate::Gate)
    push_gate!(circuit::QuantumCircuit, gates::Array{Gate})

Pushes a single gate or an array of gates to the circuit pipeline. This function is mutable.

Examples

julia> c = Snowflake.QuantumCircuit(qubit_count = 2, bit_count = 0);

julia> push_gate!(c, [hadamard(1),sigma_x(2)])
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H--
          
q[2]:--X--
          

julia> push_gate!(c, control_x(1,2))
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H----*--
            |  
q[2]:--X----X--

Snowflake.pop_gate! — Function

    pop_gate!(circuit::QuantumCircuit)

Removes the last gate from circuit.pipeline.

Examples

julia> c = Snowflake.QuantumCircuit(qubit_count = 2, bit_count = 0);

julia> push_gate!(c, [hadamard(1),sigma_x(2)])
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H--
          
q[2]:--X--
          

julia> push_gate!(c, control_x(1,2))
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H----*--
            |  
q[2]:--X----X--

julia> pop_gate!(c)
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H--
          
q[2]:--X--

Snowflake.simulate — Function

    simulate(circuit::QuantumCircuit)

Simulates and returns the wavefunction of the quantum device after running circuit.

Employs the approach described in Listing 5 of Suzuki et. al. (2021).

Examples

julia> c = Snowflake.QuantumCircuit(qubit_count = 2, bit_count = 0);

julia> push_gate!(c, hadamard(1))
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H--
          
q[2]:-----
          

julia> push_gate!(c, control_x(1,2))
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H----*--
            |  
q[2]:-------X--
               

julia> ket = simulate(c);

julia> print(ket)
4-element Ket:
0.7071067811865475 + 0.0im
0.0 + 0.0im
0.0 + 0.0im
0.7071067811865475 + 0.0im

Snowflake.simulate_shots — Function

    simulate_shots(c::QuantumCircuit, shots_count::Int = 100)

Emulates a quantum computer by running a circuit for a given number of shots and returning measurement results.

Examples

julia> c = Snowflake.QuantumCircuit(qubit_count = 2, bit_count = 0);

julia> push_gate!(c, hadamard(1))
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H--
          
q[2]:-----
          

julia> push_gate!(c, control_x(1,2))
Quantum Circuit Object:
   id: 57cf5de2-7ba7-11ec-0e10-05c6faaf91e9 
   qubit_count: 2 
   bit_count: 0 
q[1]:--H----*--
            |  
q[2]:-------X--
               

julia> simulate_shots(c, 99)
99-element Vector{String}:
 "11"
 "00"
 "11"
 "11"
 "11"
 "11"
 "11"
 "00"
 "00"
 "11"
 ⋮
 "00"
 "00"
 "11"
 "00"
 "00"
 "00"
 "00"
 "00"
 "00"

Quantum Gates

Snowflake.sigma_x — Function

sigma_x(target)

Apply the Pauli X gate to the target qubit.

The Pauli X operator is

\[\sigma_x = \begin{bmatrix} 0 & 1 \\ 1 & 0 \end{bmatrix}.\]

Snowflake.sigma_y — Function

sigma_y(target)

Apply the Pauli Y gate to the target qubit.

The Pauli Y operator is

\[\sigma_y = \begin{bmatrix} 0 & -i \\ i & 0 \end{bmatrix}.\]

Snowflake.sigma_z — Function

sigma_z(target)

Apply the Pauli Z gate to the target qubit.

The Pauli Z operator is

\[\sigma_z = \begin{bmatrix} 1 & 0 \\ 0 & -1 \end{bmatrix}.\]

Snowflake.hadamard — Function

hadamard(target)

Apply the Hadamard gate to the target qubit.

The Hadamard operator is

\[H = \frac{1}{\sqrt{2}} \begin{bmatrix} 1 & 1 \\ 1 & -1 \end{bmatrix}.\]

Snowflake.phase — Function

phase(target)

Apply the phase gate (also known as the S gate) to the target qubit.

The phase operator is

\[S = \begin{bmatrix} 1 & 0 \\ 0 & i \end{bmatrix}.\]

Snowflake.pi_8 — Function

pi_8(target)

Apply the π/8 gate (also known as the T gate) to the target qubit.

The π/8 operator is

\[T = \begin{bmatrix} 1 & 0 \\ 0 & e^{i\frac{\pi}{4}} \end{bmatrix}.\]

Snowflake.x_90 — Function

x_90(target)

Apply a 90° rotation about the X axis to the target qubit.

The corresponding operator is

\[R_x(\pi) = \begin{bmatrix} \mathrm{cos}\left(\frac{\pi}{2}\right) & -i\mathrm{sin}\left(\frac{\pi}{2}\right) \\[0.5em] -i\mathrm{sin}\left(\frac{\pi}{2}\right) & \mathrm{cos}\left(\frac{\pi}{2}\right) \end{bmatrix}.\]

Snowflake.rotation — Function

rotation(target, theta, phi)

Apply a rotation theta to the target qubit about the cos(phi)X+sin(phi)Y axis.

The corresponding operator is

\[R(\theta, \phi) = \begin{bmatrix} \mathrm{cos}\left(\frac{\theta}{2}\right) & -i e^{-i\phi} \mathrm{sin}\left(\frac{\theta}{2}\right) \\[0.5em] -i e^{i\phi} \mathrm{sin}\left(\frac{\theta}{2}\right) & \mathrm{cos}\left(\frac{\theta}{2}\right) \end{bmatrix}.\]

Snowflake.rotation_x — Function

rotation_x(target, theta)

Apply a rotation of theta about the X axis to the target qubit.

The corresponding operator is

\[R_x(\theta) = \begin{bmatrix} \mathrm{cos}\left(\frac{\theta}{2}\right) & -i\mathrm{sin}\left(\frac{\theta}{2}\right) \\[0.5em] -i\mathrm{sin}\left(\frac{\theta}{2}\right) & \mathrm{cos}\left(\frac{\theta}{2}\right) \end{bmatrix}.\]

Snowflake.rotation_y — Function

rotation_y(target, theta)

Apply a rotation of theta about the Y axis to the target qubit.

The corresponding operator is

\[R_y(\theta) = \begin{bmatrix} \mathrm{cos}\left(\frac{\theta}{2}\right) & -\mathrm{sin}\left(\frac{\theta}{2}\right) \\[0.5em] \mathrm{sin}\left(\frac{\theta}{2}\right) & \mathrm{cos}\left(\frac{\theta}{2}\right) \end{bmatrix}.\]

Snowflake.rotation_z — Function

rotation_z(target, theta)

Apply a rotation of theta about the Z axis to the target qubit.

The corresponding operator is

\[R_z(\theta) = \begin{bmatrix} \mathrm{exp}\left(-i\frac{\theta}{2}\right) & 0 \\[0.5em] 0 & \mathrm{exp}\left(i\frac{\theta}{2}\right) \end{bmatrix}.\]

Snowflake.phase_shift — Function

phase_shift(target, phi)

Apply a phase shift gate with phase shift phi to the target qubit.

The corresponding operator is

\[P(\phi) = \begin{bmatrix} i & 0 \\[0.5em] 0 & e^{i\phi} \end{bmatrix}.\]

Snowflake.universal — Function

universal(target, theta, phi, lambda)

Apply a gate which rotates the target qubit given the angles theta, phi, and lambda.

The corresponding operator is

\[U(\theta, \phi, \lambda) = \begin{bmatrix} \mathrm{cos}\left(\frac{\theta}{2}\right) & -e^{i\lambda}\mathrm{sin}\left(\frac{\theta}{2}\right) \\[0.5em] e^{i\phi}\mathrm{sin}\left(\frac{\theta}{2}\right) & e^{i\left(\phi+\lambda\right)}\mathrm{cos}\left(\frac{\theta}{2}\right) \end{bmatrix}.\]

Snowflake.control_z — Function

control_z(control_qubit, target_qubit)

Apply a controlled-Z gate given a control_qubit and a target_qubit.

The controlled-Z operator is

\[CZ = \begin{bmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 1 & 0 \\ 0 & 0 & 0 & -1 \\ \end{bmatrix}.\]

Snowflake.control_x — Function

control_x(control_qubit, target_qubit)

Apply a controlled-X gate (also known as a controlled NOT gate) given a control_qubit and a target_qubit.

The controlled-X operator is

\[CX = \begin{bmatrix} 1 & 0 & 0 & 0 \\ 0 & 1 & 0 & 0 \\ 0 & 0 & 0 & 1 \\ 0 & 0 & 1 & 0 \\ \end{bmatrix}.\]

Snowflake.iswap — Function

iswap(qubit_1, qubit_2)

Apply an imaginary swap gate to qubit_1 and qubit_2.

The imaginary swap operator is

\[i\mathrm{SWAP} = \begin{bmatrix} 1 & 0 & 0 & 0 \\ 0 & 0 & i & 0 \\ 0 & i & 0 & 0 \\ 0 & 0 & 0 & 1 \\ \end{bmatrix}.\]

Quantum Processing Unit

Snowflake.QPU — Type

Represnts a Quantum Processing Unit (QPU). Fields

manufacturer:: String – qpu manufacturer (e.g. "anyon")

generation:: String – qpu generation (e.g. "yukon")

serial_number:: String – qpu serial_number (e.g. "ANYK202201")

host:: String – the remote host url address to send the jobs to

qubit_count:: Int – number of physical qubits on the machine

connectivity::SparseArrays.SparseMatrixCSC{Int} – a matrix describing the connectivity between qubits

native_gates:: Vector{String} – the vector of native gates symbols supported by the qpu architecture

```

Snowflake.create_virtual_qpu — Function

create_virtual_qpu(qubit_count::Int, connectivity::Matrix{Int},
    native_gates::Vector{String}, host = "localhost:5600")

Creates a virtual quantum processor with qubit_count number of qubits, a connectivity matrix, and a vector of native_gates. The return value is a QPU stucture (see QPU).

Examples

To generate a QPU structure, the connectivity must be specified. Let's assume that we have a 3-qubit device where there is connectivity between qubits 2 and 1 as well as between qubits 2 and 3. If qubit 2 can only be a control qubit, the connectivity matrix corresponds to:

julia> connectivity = [1 0 0
                       1 1 1
                       0 0 1]
3×3 Matrix{Int64}:
 1  0  0
 1  1  1
 0  0  1

Here, the ones in the diagonal indicate that all qubits can perform single-qubit gates. If there is a one in an off-diagonal entry with row i and column j, it indicates that a two-qubit gate with control qubit i and target qubit j can be applied.

If the native gates are the Pauli-X gate, the Hadamard gate, and the control-X gate, the QPU can be created as follows:

julia> qpu = create_virtual_qpu(3, connectivity, ["x", "h", "cx"])
Quantum Processing Unit:
   manufacturer: none
   generation: none 
   serial_number: 00 
   host: localhost:5600 
   qubit_count: 3 
   native_gates: ["x", "h", "cx"] 
   connectivity = sparse([1, 2, 2, 2, 3], [1, 1, 2, 3, 3], [1, 1, 1, 1, 1], 3, 3)

Quantum Toolkit

Basic Quantum Objects

There are three basic quantum objects in Snowflake to simulate a quantum system. These objects are Ket, Bra, and Operator.

Snowflake.Ket — Type

A Ket represents a quantum wavefunction and is mathematically equivalent to a column vector of complex values. The norm of a Ket should always be unity.

Fields

data – the stored values.

Examples

Although NOT the preferred way, one can directly build a Ket object by passing a column vector as the initializer.

julia> using Snowflake

julia> ψ = Snowflake.Ket([1.0; 0.0; 0.0]);

julia> print(ψ)
3-element Ket:
1.0 + 0.0im
0.0 + 0.0im
0.0 + 0.0im

A better way to initialize a Ket is to use a pre-built basis such as the fock basis. See fock for further information on this function.

julia> ψ = Snowflake.fock(2, 3);

julia> print(ψ)
3-element Ket:
0.0 + 0.0im
0.0 + 0.0im
1.0 + 0.0im

Snowflake.Bra — Type

A structure representing a Bra (i.e. a row vector of complex values). A Bra is created as the complex conjugate of a Ket.

Fields

data – the stored values.

Examples

julia> ψ = Snowflake.fock(1, 3);

julia> print(ψ)
3-element Ket:
0.0 + 0.0im
1.0 + 0.0im
0.0 + 0.0im

julia> _ψ = Snowflake.Bra(ψ);

julia> print(_ψ)
Bra(Any[0.0 - 0.0im 1.0 - 0.0im 0.0 - 0.0im])

julia> _ψ * ψ    # A Bra times a Ket is a scalar
1.0 + 0.0im

julia> ψ*_ψ     # A Ket times a Bra is an operator
(3, 3)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im

Snowflake.Operator — Type

A structure representing a quantum operator (i.e. a complex matrix).

Fields

data – the complex matrix.

Examples

julia> z = Snowflake.Operator([1.0 0.0;0.0 -1.0])
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
1.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    -1.0 + 0.0im

Alternatively:

julia> z = Snowflake.sigma_z()  #sigma_z is a defined function in Snowflake
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
1.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    -1.0 + 0.0im

Base.adjoint — Function

Base.adjoint(x)

Compute the adjoint (a.k.a. conjugate transpose) of a Ket, a Bra, or an Operator.

Base.getindex — Method

Base.getindex(A::Operator, m::Int64, n::Int64)

Return the element at row m and column n of Operator A.

Snowflake.eigen — Function

eigen(A::Operator)

Compute the eigenvalue decomposition of Operator A and return an Eigen factorization object F. Eigenvalues are found in F.values while eigenvectors are found in the matrix F.vectors. Each column of this matrix corresponds to an eigenvector. The ith eigenvector is extracted by calling F.vectors[:, i].

Examples

julia> X = sigma_x()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}:
0.0 + 0.0im    1.0 + 0.0im
1.0 + 0.0im    0.0 + 0.0im

julia> F = eigen(X);

julia> eigenvalues = F.values
2-element Vector{Float64}:
 -1.0
  1.0

julia> eigenvector_1 = F.vectors[:, 1]
2-element Vector{ComplexF64}:
 -0.7071067811865475 + 0.0im
  0.7071067811865475 + 0.0im

Snowflake.tr — Function

tr(A::Operator)

Compute the trace of Operator A.

Examples

julia> I = eye()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}:
1.0 + 0.0im    0 + 0im
0 + 0im    1.0 + 0.0im

julia> trace = tr(I)
2.0 + 0.0im

Snowflake.expected_value — Function

expected_value(A::Operator, psi::Ket)

Compute the expectation value ⟨ψ|A|ψ⟩ given Operator A and Ket |ψ⟩.

Examples

julia> ψ = Ket([0.0; 1.0]);

julia> print(ψ)
2-element Ket:
0.0 + 0.0im
1.0 + 0.0im

julia> A = sigma_z()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}:
1.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    -1.0 + 0.0im

julia> expected_value(A, ψ)
-1.0 + 0.0im

Base.kron — Function

kron(x, y)

Compute the Kronecker product of two Kets or two Operators. More details about the Kronecker product can be found here.

Examples

julia> ψ_0 = Ket([0.0; 1.0]);

julia> print(ψ_0)
2-element Ket:
0.0 + 0.0im
1.0 + 0.0im

julia> ψ_1 = Ket([1.0; 0.0]);

julia> print(ψ_1)
2-element Ket:
1.0 + 0.0im
0.0 + 0.0im

julia> ψ_0_1 = kron(ψ_0, ψ_1);

julia> print(ψ_0_1)
4-element Ket:
0.0 + 0.0im
0.0 + 0.0im
1.0 + 0.0im
0.0 + 0.0im

julia> kron(sigma_x(), sigma_y())
(4, 4)-element Snowflake.Operator:
Underlying data Matrix{Complex}:
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 - 1.0im
0.0 + 0.0im    0.0 + 0.0im    0.0 + 1.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 - 1.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 1.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im

Snowflake.MultiBodySystem — Type

A structure representing a quantum multi-body system.

Fields

hilbert_space_structure – a vector of integers specifying the local Hilbert space size for each "body" within the multi-body system.

Snowflake.commute — Function

Snowflake.commute(A::Operator, B::Operator)

Returns the commutation of A and B.

julia> σ_x = sigma_x()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 0.0im    1.0 + 0.0im
1.0 + 0.0im    0.0 + 0.0im

julia> σ_y = sigma_y()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 0.0im    0.0 - 1.0im
0.0 + 1.0im    0.0 + 0.0im

julia> Snowflake.commute(σ_x,σ_y)
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 2.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 - 2.0im

Snowflake.anticommute — Function

Snowflake.anticommute(A::Operator, B::Operator)

Returns the anticommutation of A and B.

julia> σ_x = Snowflake.sigma_x()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 0.0im    1.0 + 0.0im
1.0 + 0.0im    0.0 + 0.0im

julia> Snowflake.anticommute(σ_x,σ_x)
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
2.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    2.0 + 0.0im

Snowflake.normalize! — Function

Snowflake.normalize!(x::Ket)

Normalizes Ket x such that its magnitude becomes unity.

Snowflake.ket2dm — Function

Snowflake.ket2dm(ψ)

Returns the density matrix corresponding to the pure state ψ.

Snowflake.fock_dm — Function

Snowflake.fock_dm(i, hspace_size)

Returns the density matrix corresponding to the Fock base i defined in a Hilbert space of size hspace_size.

Snowflake.moyal — Function

Snowflake.moyal(m, n)

Returns the Moyal function w_mn(eta) for Fock states m and n.

Snowflake.genlaguerre — Function

Snowflake.genlaguerre(x, alpha, n)

Returns the generalized Laguerre polynomial of degree n for x using a recursive method. See https://en.wikipedia.org/wiki/Laguerre_polynomials.

Snowflake.get_embed_operator — Function

get_embed_operator(op::Operator, target_body_index::Int, system::MultiBodySystem)

Uses a local operator (op), which is defined for a particular body (e.g. qubit) with index target_body_index, to build the corresponding operator for the Hilbert space of the multi-body system given by system.

Examples

julia> system = Snowflake.MultiBodySystem(3,2)
Snowflake.Multibody system with 3 bodies
   Hilbert space structure:
   [2, 2, 2]

julia> x = Snowflake.sigma_x()
(2, 2)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 0.0im    1.0 + 0.0im
1.0 + 0.0im    0.0 + 0.0im

julia> X_1=Snowflake.get_embed_operator(x,1,system)
(8, 8)-element Snowflake.Operator:
Underlying data Matrix{Complex}: 
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    1.0 + 0.0im
1.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im
0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    1.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im    0.0 + 0.0im

Snowflake.fock — Function

Snowflake.fock(i, hspace_size)

Returns the ith fock basis of a Hilbert space with size hspace_size as Snowflake.Ket.

Examples

julia> ψ = Snowflake.fock(0, 3);

julia> print(ψ)
3-element Ket:
1.0 + 0.0im
0.0 + 0.0im
0.0 + 0.0im

julia> ψ = Snowflake.fock(1, 3);

julia> print(ψ)
3-element Ket:
0.0 + 0.0im
1.0 + 0.0im
0.0 + 0.0im

Snowflake.create — Function

Snowflake.create(hspace_size)

Returns the bosonic creation operator for a Fock space of size hspace_size.

Snowflake.destroy — Function

Snowflake.destroy(hspace_size)

Returns the bosonic annhilation operator for a Fock space of size hspace_size.

Snowflake.number_op — Function

Snowflake.number_op(hspace_size)

Returns the number operator for a Fock space of size hspace_size.

Snowflake.coherent — Function

Snowflake.coherent(alpha, hspace_size)

Returns a coherent state for the parameter alpha in a Fock space of size hspace_size. Note that |alpha|^2 is equal to the photon number of the coherent state.

# Examples

julia> ψ = Snowflake.coherent(2.0,20);

julia> print(ψ)
20-element Ket:
0.13533528323661270231781372785917483270168304443359375 + 0.0im
0.2706705664732254046356274557183496654033660888671875 + 0.0im
0.3827859860416437253261507804308297496944779411605434060697044368322244814859633 + 0.0im
0.4420031841663186705315006220668383887063770056788388080454298547413058719111879 + 0.0im
0.4420031841663186705315006220668383887063770056788388080454298547413058719111879 + 0.0im
0.3953396664268989033516298387998153143981494385130297054512994395645417722835952 + 0.0im
0.3227934859426706749083446895240143309122789082442331409841890434072244670369041 + 0.0im
0.2440089396102658373848913914105868080225858281751344102479261185426274154783478 + 0.0im
0.1725403758685577344434702345068468523504659376126805082402433361167676595291802 + 0.0im
0.1150269172457051562956468230045645682336439584084536721601622240778451063527861 + 0.0im
0.07274941014482606043765122911007029674853133081310976424472247415659623989683902 + 0.0im
0.04386954494001140575894979175461054210856445342112420740912216424244799751166167 + 0.0im
0.02532809358034196997591593372015585816248494654573845414140041049288863525802268 + 0.0im
0.01404949847902665677216550321294394000313011924224810466364209088409881639691112 + 0.0im
0.007509772823502763531724947918871845905858490361570411398782773832262018118359266 + 0.0im
0.003878030010563633897440227516084030465660984499951448370503442999620168228954113 + 0.0im
0.001939015005281816948720113758042015232830492249975724185251721499810084114477056 + 0.0im
0.0009405604325217079112661845386949416195293171394724978755464370121167236584289665 + 0.0im
0.0004433844399679012093293182780011289711800019436937749734346315219336842860352511 + 0.0im
0.0002034387333640481868882144439914691756776463670619686354629916554722074664961924 + 0.0im

julia> Snowflake.expected_value(Snowflake.number_op(20),ψ)
3.999999793648639261230596388008292158320219007459973469036845972185905095821291 + 0.0im

Snowflake.sesolve — Function

Snowflake.sesolve(H::Operator, ψ_0::Ket, t_range::StepRangeLen; e_ops::Vector{Operator}=(Operator)[], kwargs...)
Snowflake.sesolve(H::Function, ψ_0::Ket, t_range::StepRangeLen; e_ops::Vector{Operator}=(Operator)[], kwargs...)

Solves the Shrodinger equation:

$\frac{d \Psi}{d t}=-i \hat{H}\Psi$

Fields

H – the Hamiltonian operator or a function that returns the Hamiltonian as a function of time.

ψ_0 – initital status of a quantum system

t_range – time interval for which the system has to be simulated.

e_ops – list of operators for which the expected value will be returned as a function of time.

Snowflake.mesolve — Function

Snowflake.mesolve(H::Operator, ψ_0::Ket, t_range::StepRangeLen; e_ops::Vector{Operator}=(Operator)[], kwargs...)
Snowflake.mesolve(H::Function, ψ_0::Ket, t_range::StepRangeLen; e_ops::Vector{Operator}=(Operator)[], kwargs...)

Solves the Lindblad Master equation:

$\dot{\rho}=-i [H, \rho]+\sum_i \gamma_i\left(L_i \rho L^{\dag}_i - \frac{1}{2}\left\{L^{\dag}_i L_i, \rho\right\}\right)$

Fields

H – the Hamiltonian operator or a function that returns the Hamiltonian as a function of time.

ψ_0 – initital status of a quantum system

t_range – time interval for which the system has to be simulated.

e_ops – list of operators for which the expected value will be returned as function of time.

c_ops – list of collapse operators $L_i$'s.