Quantum Processing Unit

Snowflurry.AnyonYukonQPU — Type

AnyonYukonQPU <: AbstractQPU

A data structure to represent an Anyon System's Yukon generation QPU, consisting of 6 qubits in a linear arrangement (see LineConnectivity).

Fields

client ::Client – Client to the QPU server.

status_request_throttle ::Function – Used to rate-limit job status requests.

project_id ::String – Used to identify which project the jobs sent to this QPU belong to.

Example

julia>  qpu = AnyonYukonQPU(host = "example.anyonsys.com", user = "test_user", access_token = "not_a_real_access_token", project_id = "9d6949c8-bb5d-4aeb-9aa3-e7b284f0f269")
Quantum Processing Unit:
   manufacturer:  Anyon Systems Inc.
   generation:    Yukon
   serial_number: ANYK202201
   project_id:    9d6949c8-bb5d-4aeb-9aa3-e7b284f0f269
   qubit_count:   6 
   connectivity_type:  linear

Snowflurry.AnyonYamaskaQPU — Type

AnyonYamaskaQPU <: AbstractQPU

A data structure to represent an Anyon System's Yamaska generation QPU, consisting of 12 qubits in a 2D lattice arrangement (see LatticeConnectivity).

Fields

client ::Client – Client to the QPU server.

status_request_throttle ::Function – Used to rate-limit job status requests.

project_id ::String – Used to identify which project the jobs sent to this QPU belong to.

Example

julia>  qpu = AnyonYamaskaQPU(host = "example.anyonsys.com", user = "test_user", access_token = "not_a_real_access_token", project_id = "9d6949c8-bb5d-4aeb-9aa3-e7b284f0f269")
Quantum Processing Unit:
   manufacturer:  Anyon Systems Inc.
   generation:    Yamaska
   serial_number: ANYK202301
   project_id:    9d6949c8-bb5d-4aeb-9aa3-e7b284f0f269
   qubit_count:   12 
   connectivity_type:  2D-lattice

Snowflurry.VirtualQPU — Type

VirtualQPU

A data structure to represent a Quantum Simulator.

Example

julia> qpu = VirtualQPU()
Quantum Simulator:
   developers:  Anyon Systems Inc.
   package:     Snowflurry.jl

Snowflurry.Client — Type

Client

A data structure to represent a Client to a QPU service.

Fields

host::String – URL of the QPU server.

user::String – Username.

access_token::String – User access token.

Example

julia> c = Client(host = "http://example.anyonsys.com", user = "test_user", access_token = "not_a_real_access_token")
Client for QPU service:
   host:         http://example.anyonsys.com
   user:         test_user 
 

Snowflurry.get_host — Function

get_host(Client)

Returns host URL of a Client to a QPU service.

Example

julia> c = Client(host = "http://example.anyonsys.com", user = "test_user", access_token = "not_a_real_access_token");

julia> get_host(c)
"http://example.anyonsys.com"

Snowflurry.submit_job — Function

submit_job(client::Client,circuit::QuantumCircuit,shot_count::Integer)

Submit a circuit to a Client of QPU service, requesting a number of repetitions (shot_count). Returns circuitID.

Example

julia> submit_job(client, QuantumCircuit(qubit_count = 3, instructions = [sigma_x(3), control_z(2, 1), readout(1, 1)]), 100, "project_id")
"8050e1ed-5e4c-4089-ab53-cccda1658cd0"

Snowflurry.get_status — Function

get_status(client::Client,circuitID::String)::Tuple{Status,Dict{String,Int}}

Obtain the status of a circuit computation through a Client of a QPU service. Returns status::Dict containing status["type"]: -"QUEUED" : Computation in queue -"RUNNING" : Computation being processed -"FAILED" : QPU service has returned an error message -"SUCCEEDED": Computation is completed, result is available.

In the case of status["type"]=="FAILED", the server error is contained in status["message"].

In the case of status["type"]=="SUCCEEDED", the second element in the return Tuple is the histogram of the job results, as computed on the QPU.

Example

julia> jobID = submit_job(client, QuantumCircuit(qubit_count = 3, instructions = [sigma_x(3), control_z(2, 1), readout(1, 1)]), 100, "project_id")
"8050e1ed-5e4c-4089-ab53-cccda1658cd0"

julia> get_status(client, jobID)
(Status: SUCCEEDED
, Dict("001" => 100))

Snowflurry.run_job — Function

run_job(qpu::VirtualQPU, circuit::QuantumCircuit, shot_count::Integer)

Run a circuit computation on a QPU simulator, repeatedly for the specified number of repetitions (shot_count). Returns the histogram of the completed circuit measurements, as prescribed by the Readouts present.

Example

julia> qpu = VirtualQPU();

julia> run_job(qpu, QuantumCircuit(qubit_count = 3, instructions = [sigma_x(3), control_z(2, 1), readout(1, 1), readout(2, 2), readout(3, 3)]), 100)
Dict{String, Int64} with 1 entry:
  "001" => 100

run_job(qpu::AnyonYukonQPU, circuit::QuantumCircuit, shot_count::Integer)

Run a circuit computation on a QPU service, repeatedly for the specified number of repetitions (shot_count). Returns the histogram of the completed circuit calculations, or an error message. If the circuit received in invalid - for instance, it is missing a Readout - it is not sent to the host, and an error is throw.

Example

julia> qpu = AnyonYukonQPU(client, "project_id");

julia> run_job(qpu, QuantumCircuit(qubit_count = 3, instructions = [sigma_x(3), control_z(2, 1), readout(1, 1)]), 100)
Dict{String, Int64} with 1 entry:
  "001" => 100

Snowflurry.transpile_and_run_job — Function

transpile_and_run_job(qpu::VirtualQPU, circuit::QuantumCircuit,shot_count::Integer; transpiler::Transpiler = get_transpiler(qpu))

This method first transpiles the input circuit using either the default transpiler, or any other transpiler passed as a key-word argument. The transpiled circuit is then run on a QPU simulator, repeatedly for the specified number of repetitions (shot_count). Returns the histogram of the completed circuit calculations, or an error message.

Example

julia> qpu=VirtualQPU();

julia> transpile_and_run_job(qpu, QuantumCircuit(qubit_count = 3, instructions = [sigma_x(3), control_z(2, 1), readout(1, 1), readout(2, 2), readout(3, 3)]) ,100)
Dict{String, Int64} with 1 entry:
  "001" => 100

transpile_and_run_job(qpu::AnyonYukonQPU, circuit::QuantumCircuit,shot_count::Integer;transpiler::Transpiler=get_transpiler(qpu))

This method first transpiles the input circuit using either the default transpiler, or any other transpiler passed as a key-word argument. The transpiled circuit is then run on the AnyonYukonQPU, repeatedly for the specified number of repetitions (shot_count).

Returns the histogram of the completed circuit calculations, or an error message.

Example

julia> qpu = AnyonYukonQPU(client_anyon, "project_id");

julia> transpile_and_run_job(qpu, QuantumCircuit(qubit_count = 3, instructions = [sigma_x(3), control_z(2, 1), readout(3, 3)]), 100)
Dict{String, Int64} with 1 entry:
  "001" => 100

Snowflurry.get_transpiler — Function

get_transpiler(qpu::AbstractQPU)::Transpiler

Returns the transpiler associated with this QPU.

Example

julia> qpu = AnyonYukonQPU(client, "project_id");

julia> get_transpiler(qpu)
SequentialTranspiler(Transpiler[CircuitContainsAReadoutTranspiler(), ReadoutsDoNotConflictTranspiler(), UnsupportedGatesTranspiler(), DecomposeSingleTargetSingleControlGatesTranspiler(), CastToffoliToCXGateTranspiler(), CastCXToCZGateTranspiler(), CastISwapToCZGateTranspiler(), SwapQubitsForAdjacencyTranspiler(LineConnectivity{6}
1──2──3──4──5──6
), CastSwapToCZGateTranspiler(), CompressSingleQubitGatesTranspiler(), SimplifyTrivialGatesTranspiler(1.0e-6), CastUniversalToRzRxRzTranspiler(), SimplifyRxGatesTranspiler(1.0e-6), CastRxToRzAndHalfRotationXTranspiler(), CompressRzGatesTranspiler(), SimplifyRzGatesTranspiler(1.0e-6), ReadoutsAreFinalInstructionsTranspiler()])

Snowflurry.SequentialTranspiler — Type

SequentialTranspiler(Vector{<:Transpiler})

Composite transpiler object which is constructed from an array of Transpiler stages. Calling transpile(::SequentialTranspiler,::QuantumCircuit) will apply each stage in sequence to the input circuit and return a transpiled output circuit. The result of the input and output circuit on any arbitrary state Ket is unchanged (up to a global phase).

Examples

julia> transpiler = SequentialTranspiler([CompressSingleQubitGatesTranspiler(), CastToPhaseShiftAndHalfRotationXTranspiler()]);

julia> circuit = QuantumCircuit(qubit_count = 2, instructions = [sigma_x(1), hadamard(1)])
Quantum Circuit Object:
   qubit_count: 2 
   bit_count: 2 
q[1]:──X────H──
               
q[2]:──────────
               

julia> transpile(transpiler,circuit)
Quantum Circuit Object:
   qubit_count: 2 
   bit_count: 2 
q[1]:──Z────X_90────Z_90────X_m90────Z──
                                                              
q[2]:───────────────────────────────────
                                                              

julia> circuit = QuantumCircuit(qubit_count = 3, instructions = [sigma_x(1),sigma_y(1),control_x(2,3),phase_shift(1,π/3)])
Quantum Circuit Object:
   qubit_count: 3 
   bit_count: 3 
q[1]:──X────Y─────────P(1.0472)──

q[2]:────────────*───────────────
                 |               
q[3]:────────────X───────────────
                                 

julia> transpile(transpiler,circuit)
Quantum Circuit Object:
   qubit_count: 3 
   bit_count: 3 
q[1]:──P(-2.0944)───────
                        
q[2]:────────────────*──
                     |  
q[3]:────────────────X──
                        

Snowflurry.AllToAllConnectivity — Type

AllToAllConnectivity <:AbstractConnectivity

A data structure to represent all-to-all qubit connectivity in an Anyon System's QPU. This connectivity type is encountered in simulated QPUs, such as the VirtualQPU

Example

julia> connectivity = AllToAllConnectivity()
AllToAllConnectivity()

Snowflurry.LineConnectivity — Type

LineConnectivity <:AbstractConnectivity

A data structure to represent linear qubit connectivity in an Anyon System's QPU. This connectivity type is encountered in QPUs such as the AnyonYukonQPU

Fields

dimension ::Int – Qubit count in this connectivity.

Example

julia> connectivity = LineConnectivity(6)
LineConnectivity{6}
1──2──3──4──5──6

Snowflurry.LatticeConnectivity — Type

LatticeConnectivity <:AbstractConnectivity

A data structure to represent 2D-lattice qubit connectivity in an Anyon System's QPU. This connectivity type is encountered in QPUs such as the AnyonYamaskaQPU

Fields

qubits_per_row ::Vector{Int} – number of qubits in each line, when constructing the printout.

dimensions ::Vector{Int} – number of rows and columns (turned 45° in the printout).

Example

The following lattice has 4 rows, made of qubits [2, 6, 10], [1, 5, 9], [4, 8, 12] and [3, 7, 11], with each of those rows having 3 columns.

The corresponding qubits_per_row field is [2,4,4,2], the number of qubits in each line in the printed representation.

julia> connectivity = LatticeConnectivity(4, 3)
LatticeConnectivity{4,3}
        1 ──  2 
        |     | 
  3 ──  4 ──  5 ──  6 
        |     |     | 
        7 ──  8 ──  9 ── 10 
              |     | 
             11 ── 12 

Lattices of arbitrary dimensions can be built:

julia> connectivity = LatticeConnectivity(6, 4)
LatticeConnectivity{6,4}
              1 ──  2 
              |     | 
        3 ──  4 ──  5 ──  6 
        |     |     |     | 
  7 ──  8 ──  9 ── 10 ── 11 ── 12 
        |     |     |     |     | 
       13 ── 14 ── 15 ── 16 ── 17 ── 18 
              |     |     |     | 
             19 ── 20 ── 21 ── 22 
                    |     | 
                   23 ── 24 

Non-rectangular shapes can also be achieved by directly specifying the desired qubits_per_row as a Vector{Int}:

julia> connectivity = LatticeConnectivity([5, 7, 8, 8, 7, 5])
LatticeConnectivity{8,5}
        1 ──  2 ──  3 ──  4 ──  5 
        |     |     |     |     | 
  6 ──  7 ──  8 ──  9 ── 10 ── 11 ── 12 
  |     |     |     |     |     |     | 
 13 ── 14 ── 15 ── 16 ── 17 ── 18 ── 19 ── 20 
        |     |     |     |     |     |     | 
       21 ── 22 ── 23 ── 24 ── 25 ── 26 ── 27 ── 28 
              |     |     |     |     |     |     | 
             29 ── 30 ── 31 ── 32 ── 33 ── 34 ── 35 
                    |     |     |     |     | 
                   36 ── 37 ── 38 ── 39 ── 40 

Snowflurry.path_search — Function

path_search(origin::Int, target::Int, connectivity::AbstractConnectivity)

Find the shortest path between origin and target qubits in terms of Manhattan distance, using the Breadth-First Search algorithm, on any connectivity::AbstractConnectivity.

Example

julia> connectivity = LineConnectivity(6)
LineConnectivity{6}
1──2──3──4──5──6

julia> path = path_search(2, 5, connectivity)
4-element Vector{Int64}:
 5
 4
 3
 2

On LatticeConnectivity, the print_connectivity() method is used to visualize the path. The qubits along the path between origin and target are marker with ( )

julia> connectivity = LatticeConnectivity(6, 4)
LatticeConnectivity{6,4}
              1 ──  2 
              |     | 
        3 ──  4 ──  5 ──  6 
        |     |     |     | 
  7 ──  8 ──  9 ── 10 ── 11 ── 12 
        |     |     |     |     | 
       13 ── 14 ── 15 ── 16 ── 17 ── 18 
              |     |     |     | 
             19 ── 20 ── 21 ── 22 
                    |     | 
                   23 ── 24 

julia> path = path_search(3, 24, connectivity)
8-element Vector{Int64}:
 24
 23
 20
 19
 14
 13
  8
  3

Snowflurry.get_adjacency_list — Function

get_adjacency_list(connectivity::AbstractConnectivity)::Dict{Int,Vector{Int}}

Given an object of type AbstractConnectivity, get_adjacency_list returns a Dict where key => value pairs are each qubit number => an Vector of the qubits that are adjacent (neighbors) to it on this particular connectivity.

Example

julia> connectivity = LineConnectivity(6)
LineConnectivity{6}
1──2──3──4──5──6

julia> get_adjacency_list(connectivity)
Dict{Int64, Vector{Int64}} with 6 entries:
  5 => [4, 6]
  4 => [3, 5]
  6 => [5]
  2 => [1, 3]
  3 => [2, 4]
  1 => [2]

julia> connectivity = LatticeConnectivity(4, 3)
LatticeConnectivity{4,3}
        1 ──  2 
        |     | 
  3 ──  4 ──  5 ──  6 
        |     |     | 
        7 ──  8 ──  9 ── 10 
              |     | 
             11 ── 12 

  
julia> get_adjacency_list(connectivity)
Dict{Int64, Vector{Int64}} with 12 entries:
  5  => [2, 8, 4, 6]
  12 => [9, 11]
  8  => [5, 11, 7, 9]
  1  => [4, 2]
  6  => [9, 5]
  11 => [8, 12]
  9  => [6, 12, 8, 10]
  3  => [4]
  7  => [4, 8]
  4  => [1, 7, 3, 5]
  2  => [5, 1]
  10 => [9]

Note

get_adjacency_list cannot be performed for AllToAllConnectivity, as in such a connectivity, all qubits are adjacent, with no upper bound on the number of qubits. A finite list of adjacent qubits thus cannot be constructed.

Snowflurry.get_qubits_distance — Function

get_qubits_distance(target_1::Int, target_2::Int, ::AbstractConnectivity)

Find the length of the shortest path between target qubits in terms of Manhattan distance, using the Breadth-First Search algorithm, on any connectivity::AbstractConnectivity.

Example

julia>  connectivity = LineConnectivity(6)
LineConnectivity{6}
1──2──3──4──5──6

julia> get_qubits_distance(2, 5, connectivity)
3

julia> connectivity = LatticeConnectivity(6, 4)
LatticeConnectivity{6,4}
              1 ──  2 
              |     | 
        3 ──  4 ──  5 ──  6 
        |     |     |     | 
  7 ──  8 ──  9 ── 10 ── 11 ── 12 
        |     |     |     |     | 
       13 ── 14 ── 15 ── 16 ── 17 ── 18 
              |     |     |     | 
             19 ── 20 ── 21 ── 22 
                    |     | 
                   23 ── 24 

julia> get_qubits_distance(3, 24, connectivity)
7